+++ packages/examples/packages/bip32/src/index.tsx
@@ -1,14 +1,11 @@
 import type { OnRpcRequestHandler } from '@metamask/snaps-sdk';
 import {
   DialogType,
-  panel,
-  text,
-  heading,
-  copyable,
   InvalidParamsError,
   UserRejectedRequestError,
   MethodNotFoundError,
 } from '@metamask/snaps-sdk';
+import { Box, Copyable, Heading, Text } from '@metamask/snaps-sdk/jsx';
 import {
   add0x,
   assert,
@@ -30,6 +27,7 @@ import { getPrivateNode, getPublicKey } from './utils';
  * key is returned in hex format.
  * - `signMessage`: Derive a BIP-32 private key for a given BIP-32 path, and use
  * it to sign a message. The signature is returned in hex format.
+ * - `getEntropySources`: Get the list of entropy sources available to the Snap.
  *
  * @param params - The request parameters.
  * @param params.request - The JSON-RPC request object.
@@ -62,13 +60,16 @@ export const onRpcRequest: OnRpcRequestHandler = async ({ request }) => {
         method: 'snap_dialog',
         params: {
           type: DialogType.Confirmation,
-          content: panel([
-            heading('Signature request'),
-            text(
-              `Do you want to ${curve} sign "${message}" with the following public key?`,
-            ),
-            copyable(add0x(node.publicKey)),
-          ]),
+          content: (
+            <Box>
+              <Heading>Signature request</Heading>
+              <Text>
+                Do you want to {curve} sign "{message}" with the following
+                public key?
+              </Text>
+              <Copyable value={add0x(node.publicKey)} />
+            </Box>
+          ),
         },
       });
 
@@ -95,6 +96,12 @@ export const onRpcRequest: OnRpcRequestHandler = async ({ request }) => {
       // This is guaranteed to never happen because of the `assert` above. But
       // TypeScript doesn't know that, so we need to throw an error here.
       throw new Error(`Unsupported curve: ${String(curve)}.`);
+    }
+
+    case 'getEntropySources': {
+      return await snap.request({
+        method: 'snap_listEntropySources',
+      });
     }
 
     default: