+++ packages/snaps-rpc-methods/src/restricted/getBip32Entropy.ts
@@ -16,15 +16,20 @@ import type { NonEmptyArray } from '@metamask/utils';
 import { assert } from '@metamask/utils';
 
 import type { MethodHooksObject } from '../utils';
-import { getNode } from '../utils';
+import { getSecretRecoveryPhrase, getNode } from '../utils';
 
 const targetName = 'snap_getBip32Entropy';
 
 export type GetBip32EntropyMethodHooks = {
   /**
-   * @returns The mnemonic of the user's primary keyring.
+   * Get the mnemonic of the provided source. If no source is provided, the
+   * mnemonic of the primary keyring will be returned.
+   *
+   * @param source - The optional ID of the source to get the mnemonic of.
+   * @returns The mnemonic of the provided source, or the default source if no
+   * source is provided.
    */
-  getMnemonic: () => Promise<Uint8Array>;
+  getMnemonic: (source?: string | undefined) => Promise<Uint8Array>;
 
   /**
    * Waits for the extension to be unlocked.
@@ -125,10 +130,15 @@ export function getBip32EntropyImplementation({
     const { params } = args;
     assert(params);
 
+    const secretRecoveryPhrase = await getSecretRecoveryPhrase(
+      getMnemonic,
+      params.source,
+    );
+
     const node = await getNode({
       curve: params.curve,
       path: params.path,
-      secretRecoveryPhrase: await getMnemonic(),
+      secretRecoveryPhrase,
       cryptographicFunctions: getClientCryptography(),
     });