+++ packages/snaps-sdk/src/types/methods/get-bip32-public-key.ts
@@ -2,14 +2,19 @@ import type { Bip32Entropy } from '../permissions';
 
 /**
  * The request parameters for the `snap_getBip32PublicKey` method.
- *
- * @property path - The BIP-32 path to derive the public key from.
- * @property curve - The curve to use when deriving the public key.
- * @property compressed - Whether to return the compressed public key. Defaults
- * to `false`.
  */
 export type GetBip32PublicKeyParams = Bip32Entropy & {
+  /**
+   * Whether to return the compressed public key. Defaults to `false`.
+   */
   compressed?: boolean;
+
+  /**
+   * The ID of the entropy source to use. If not specified, the primary entropy
+   * source will be used. For a list of available entropy sources, see the
+   * `snap_listEntropySources` method.
+   */
+  source?: string | undefined;
 };
 
 /**