+++ scripts/install-chrome.sh
@@ -5,12 +5,12 @@ set -u
 set -o pipefail
 
 # To get the latest version, run `yarn update-chrome`
-CHROME_VERSION='133.0.6943.141-1'
+CHROME_VERSION='134.0.6998.35-1'
 CHROME_BINARY="google-chrome-stable_${CHROME_VERSION}_amd64.deb"
 CHROME_BINARY_URL="https://dl.google.com/linux/chrome/deb/pool/main/g/google-chrome-stable/${CHROME_BINARY}"
 
 # To retrieve this checksum, run the `wget` and `shasum` commands below
-CHROME_BINARY_SHA512SUM='84aad5d0f5ebd717ed8baf781e19be3c956503c0c13c4ed7b9e6e86194b2d682b8de8a1dbdb62c09b59d22b626655242046167f51a73dd40ac771e2db3aa26c7'
+CHROME_BINARY_SHA512SUM='acb6ce4a2c02e5c3db6ef067d2aab62b8097f05370c96897c9b2605e344afba6dc3f35e8def5e7370eeb9d866139c25bdd6e541256aa9b1cf679e46d58033cb6'
 
 wget -O "${CHROME_BINARY}" -t 5 "${CHROME_BINARY_URL}"